AnimalsPage addition
