+++ backend/app/__main__.py
@@ -1,7 +1,6 @@
 # Entry point for `python -m app`, import create_app from __init__ and run the dev server
 from app import create_app
 
-app = create_app()
-
 if __name__ == "__main__":
-    app.run(debug=True, port=8081)
+    app = create_app()
+    app.run(debug=True, host='0.0.0.0', port=8081)
